// src/dimension/TimeSlotPeriodicity.enum.ts
export enum TimeSlotPeriodicity {
  Day = 'day',
  MonthWeekSat = 'month_week_sat',
  MonthWeekSun = 'month_week_sun',
  MonthWeekMon = 'month_week_mon',
  WeekSat = 'week_sat',
  WeekSun = 'week_sun',
  WeekMon = 'week_mon',
  Month = 'month',
  Quarter = 'quarter',
  Semester = 'semester',
  Year = 'year',
  All = 'all',
}